pyproject.toml
[project]
name = "libra"
version = "0.1.0"
description = "High-performance AI trading platform with LMAX Disruptor-inspired message bus"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
authors = [
    {name = "Libra Team"}
]
keywords = ["trading", "algorithmic-trading", "message-bus", "async", "high-performance"]

dependencies = [
    "msgspec>=0.18",                           # Fast structs (4x faster than dataclass)
    "aiohttp>=3.9",                            # Async HTTP client
    "rich>=13.0",                              # Pretty terminal output
    "orjson>=3.10",                            # 10x faster JSON serialization
    "uvloop>=0.21; sys_platform != 'win32'",   # 2-4x faster asyncio (Unix only)
    "polars>=1.0",                             # 10-3500x faster than pandas
    "pyyaml>=6.0",                             # YAML config files
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-benchmark>=4.0",
    "pytest-cov>=4.0",
    "pytest-xdist>=3.5",   # Parallel test execution
    # Benchmarking
    "hdrhistogram>=0.10",  # High-precision latency histograms
    # Linting & Formatting
    "ruff>=0.8",
    "mypy>=1.8",
    # Pre-commit hooks
    "pre-commit>=3.6",
]

# Exchange connectivity plugins (Issue #29)
ccxt = [
    "ccxt>=4.0",
]

# All optional plugins (convenience install)
all-plugins = [
    "libra[ccxt,freqtrade,tui,database,api]",
]

# TUI - Terminal User Interface (v4.0+ required for Markdown streaming)
tui = [
    "textual>=7.0.0",       # v7.0 adds atom-one themes, v6.0 adds TextArea suggestions
    "textual-plotext>=1.0", # Plotext charts in Textual (Issue #40)
]

# Future: validation (only at API boundaries)
validation = [
    "pydantic>=2.0",
]

# Database - QuestDB time-series database (Issue #21)
database = [
    "questdb>=3.0",      # QuestDB Python client (ILP ingestion)
    "asyncpg>=0.29",     # Async PostgreSQL driver (queries via PGWire)
]

# Freqtrade - Strategy adapter plugin (Issue #6 / Phase 2.1)
freqtrade = [
    "freqtrade>=2024.1",  # Freqtrade trading bot
    "technical>=1.4",      # Freqtrade indicator library
    "pandas>=2.0",         # Required for DataFrames
    "pandas-ta>=0.3.14b",  # Additional technical indicators
]

# REST API - FastAPI endpoints (Issue #30)
api = [
    "fastapi>=0.115",      # Modern web framework
    "uvicorn[standard]>=0.32",  # ASGI server with uvloop
    "python-jose[cryptography]>=3.3",  # JWT authentication
    "passlib[bcrypt]>=1.7",  # Password hashing
    "python-multipart>=0.0.9",  # Form parsing
]

[project.scripts]
libra = "libra.cli:main"

# Plugin entry points for strategy discovery (ADR-007)
[project.entry-points."libra.strategies"]
freqtrade = "libra.plugins.freqtrade_adapter:FreqtradeAdapter"

# Gateway plugin entry points (Issue #29)
[project.entry-points."libra.gateways"]
paper = "libra.gateways.paper_gateway:PaperGateway"
ccxt = "libra.gateways.ccxt_gateway:CCXTGateway"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/libra"]

# =============================================================================
# pytest
# =============================================================================
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "-v",
    "--strict-markers",
    "--strict-config",
    "-ra",  # Show extra test summary
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "benchmark: marks tests as benchmarks",
    "authenticated: marks tests requiring API keys (deselect with '-m \"not authenticated\"')",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore::pytest.PytestUnraisableExceptionWarning",  # Async cleanup warnings
]

# =============================================================================
# pytest-benchmark
# =============================================================================
[tool.pytest-benchmark]
disable_gc = true
warmup = true
warmup_iterations = 1000
min_rounds = 30
max_time = 5.0
timer = "time.perf_counter_ns"

# =============================================================================
# ruff - Fast Python linter & formatter
# =============================================================================
[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "S",      # flake8-bandit (security)
    "A",      # flake8-builtins
    "COM",    # flake8-commas
    "DTZ",    # flake8-datetimez
    "T10",    # flake8-debugger
    "ISC",    # flake8-implicit-str-concat
    "ICN",    # flake8-import-conventions
    "PIE",    # flake8-pie
    "PT",     # flake8-pytest-style
    "Q",      # flake8-quotes
    "RSE",    # flake8-raise
    "RET",    # flake8-return
    "TCH",    # flake8-type-checking
    "ARG",    # flake8-unused-arguments
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented-out code)
    "PL",     # Pylint
    "TRY",    # tryceratops
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",    # Line too long (handled by formatter)
    "S101",    # Use of assert (fine in tests)
    "PLR0913", # Too many arguments
    "TRY003",  # Long exception messages
    "COM812",  # Trailing comma (conflicts with formatter)
    "ISC001",  # Implicit string concat (conflicts with formatter)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",    # assert is fine in tests
    "ARG001",  # Unused function args (fixtures)
    "PLR2004", # Magic values in tests are fine
]
"benchmarks/**/*.py" = [
    "S101",      # assert is fine in benchmarks
    "PLR2004",   # Magic values in benchmarks are fine
    "ARG001",    # Unused function args (handler signatures)
    "ARG002",    # Unused method args (fixtures for side-effects)
    "PLR0915",   # Too many statements (benchmark summaries)
]

[tool.ruff.lint.isort]
known-first-party = ["libra"]
force-single-line = false
lines-after-imports = 2

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true

# =============================================================================
# mypy - Static type checker
# =============================================================================
[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
warn_unreachable = true
show_error_codes = true
show_column_numbers = true
pretty = true

# Per-module overrides
[[tool.mypy.overrides]]
module = [
    "msgspec.*",
    "pytest.*",
    "pytest_asyncio.*",
    "pytest_benchmark.*",
    "orjson.*",
    "uvloop.*",
    "polars.*",
    "questdb.*",
    "asyncpg.*",
]
ignore_missing_imports = true

# =============================================================================
# coverage
# =============================================================================
[tool.coverage.run]
source = ["src/libra"]
branch = true
parallel = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "raise AssertionError",
    "@abstractmethod",
    "assert_never",
]
fail_under = 80
show_missing = true
